Crouch Animation Altered, Original Walk had to be readded to fix issues atm.  Still issues with jump animation
The crouch animation has been heavily altered and now appears to be working correctly however this was only once the original Unity walk animation was set as the primary walk.  There are also still collider issues when the player jumps and I am still trying to determine which animation is causing the interference.